read today food and nutrient from home screen (#16)

+++ app/src/main/java/com/example/dieter/data/source/domain/NutrientModel.kt
@@ -2,7 +2,7 @@ package com.example.dieter.data.source.domain
 
 data class NutrientModel(
     val name: String,
-    val current: Int,
+    val current: Float,
     val of: Int,
     val unit: String
 )
